refactor: create widgets folder for header
