+++ src/widgets/header/ui/assets/calendar-image.png
